Made sure that the listener count has actually changed when removing a once event. This stops it returning false and rolling the loop back when an item hasn't been removed.
This lack of removal happens when you call `removeAllListeners` from inside a once event. It goes looking to remove it's self but it already has done, so it does not remove anything. Yet the loop thinks something was removed so it rolls back. Thus creating an infinite loop.

What a fun bug.

Fixes issue #26