Changed the coding style to match my new one.
It's a style I have slowly learned over the years of developing
EventEmitter, so I thought I should update it with my learnings and
opinions eventually. I admit, it's influenced slighly by the rules
imposed by Amazon, but I think it's a bit cleaner now. Even if this
commit does count as deleting the whole file and rewriting it.