Added the regex event matching code. #39
This is for the getListeners and removeEvent method. By adding it to
getListeners it has added support to emitEvent as a side effect. Passing
a regex to other methods, such as removeListener will currently cause
all of reality to break down, condemning the world to thousands of years
of sorrow. Don't do it.

I'm going to try and restructure some stuff so you can use a regex for
pretty much every method. I basically need to slap square braces on a
few things and add the word "for" a few times.

Let's see what I can do.