Removed all references to the dist directory.
No longer need to output extra copies of the script, it was meant for
the downloads page of the repo. Sadly, GitHub have discontinued the
downloads page, so now all I need is a minified version at the root of
the repo. Even though I hate committing minified code.

This also removes all references to the downloads page. Fixes #41.