Named the returned alias function.
Just makes it nice and easy to see what function is what. You don't want
your application to get into a state where you have layers and layers of
unnamed functions. It's really difficult to work out where you are in
the execution. At least naming these closures gives you an idea as to
where you are in the stack.